Move loadBlacklistFile
